# golangapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ICMON project",
        "title": "Go IoT API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/mqtt/devicecontrol": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Publishes a control message to a topic (usually ending with CONTROL), then listens on the corresponding DATA topic (CONTROL replaced by DATA) for a response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Send control command to an MQTT device and wait for response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MQTT topic to send control message (e.g., BAACTW02/CONTROL)",
                        "name": "topic",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Control message (e.g., ON, OFF, 1, 0)",
                        "name": "message",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.DeviceControlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/gettopicdata": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Subscribes to the given topic, waits for a single message, then unsubscribes. Caches the result for 60 seconds.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Get live or cached MQTT topic data (request‑response)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MQTT topic name (e.g. BAACTW05/DATA)",
                        "name": "topic",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_presenter.MQTTTopicDataResponse"
                        }
                    },
                    "400": {
                        "description": "missing topic parameter",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    },
                    "500": {
                        "description": "MQTT timeout or internal error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Publishes a payload to a specified MQTT topic. Supports QoS 0,1,2 and retained flag.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Publish an MQTT message",
                "parameters": [
                    {
                        "description": "Publish request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.PublishRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.PublishResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current connection state of the MQTT client to the broker, along with a timestamp (Asia/Bangkok).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Get MQTT connection status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/subscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Subscribes to a topic and starts receiving messages (persistent subscription). The connection will stay open and messages will be broadcast via WebSocket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Subscribe to an MQTT topic",
                "parameters": [
                    {
                        "description": "Subscribe request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.SubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.SubscribeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/subscriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an array of topics that the MQTT client is currently subscribed to (persistent subscriptions).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "List all subscribed topics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.SubscriptionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
        "/mqtt/unsubscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a previous subscription from a topic. Messages will no longer be received.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mqtt"
                ],
                "summary": "Unsubscribe from an MQTT topic",
                "parameters": [
                    {
                        "description": "Unsubscribe request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenter.UnsubscribeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.UnsubscribeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/icmongolang_internal_modules_mqtt_delivery_http.errResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "icmongolang_internal_modules_mqtt_delivery_http.errResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "icmongolang_internal_modules_mqtt_presenter.MQTTTopicDataResponse": {
            "type": "object",
            "properties": {
                "cache_enabled": {
                    "type": "boolean",
                    "example": true
                },
                "cache_hit": {
                    "type": "boolean",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data_length": {
                    "type": "integer",
                    "example": 128
                },
                "error_detail": {
                    "type": "string",
                    "example": "timeout waiting for message"
                },
                "fetch_duration_ms": {
                    "type": "integer",
                    "example": 145
                },
                "from": {
                    "type": "string",
                    "example": "mqtt"
                },
                "message": {
                    "type": "string",
                    "example": "live data"
                },
                "mqtt_connected": {
                    "type": "boolean",
                    "example": true
                },
                "payload": {
                    "description": "✅ remove example:\"25.5,60,1013\""
                },
                "statuscode": {
                    "type": "integer",
                    "example": 200
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-06-08T12:36:58+07:00"
                },
                "topic": {
                    "type": "string",
                    "example": "AIRCOM4/DATA"
                }
            }
        },
...
        "presenter.DeviceControlResponse": {
            "type": "object",
            "properties": {
                "cachelift": {},
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error_detail": {
                    "type": "string"
                },
                "fetch_duration_ms": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "message_sent": {
                    "type": "string"
                },
                "message_th": {
                    "type": "string"
                },
                "payload": {},
                "status": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                },
                "statuscode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "topic_control": {
                    "type": "string"
                },
                "topic_data": {
                    "type": "string"
                }
            }
        },
...
        "presenter.PublishRequest": {
            "type": "object",
            "required": [
                "payload",
                "topic"
            ],
            "properties": {
                "payload": {},
                "qos": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                },
                "retained": {
                    "type": "boolean"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "presenter.PublishResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "presenter.SubscribeRequest": {
            "type": "object",
            "required": [
                "topic"
            ],
            "properties": {
                "qos": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "presenter.SubscribeResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "qos": {
                    "type": "integer"
                },
                "subscribed_at": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "presenter.SubscriptionsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "presenter.UnsubscribeRequest": {
            "type": "object",
            "required": [
                "topic"
            ],
            "properties": {
                "topic": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your access token. Example: \"Bearer eyJhbGciOiJSUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/auth/login",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}